hrspace_backend/requirements.txt
# requests==2.26.0
# asgiref==3.2.10
Django==3.2
# django-filter==2.4.0
djangorestframework==3.12.4
# PyJWT==2.1.0
# pytest==6.2.4
# pytest-django==4.4.0
# pytest-pythonpath==0.7.3
# djangorestframework-simplejwt==4.7.2
# django-filter==22.1
tqdm==4.66.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
django-cors-headers
drf-yasg==1.21.7